# api/openapi.json
{
  "components": {
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "The error envelope.",
        "headers": {
          "X-Request-Id": {
            "description": "Echo of the caller's id, else a new uuid4.",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "BriefOut": {
        "examples": [
          {
            "atmist": {
              "age": "58",
              "injuries": "unknown",
              "mechanism": "unknown",
              "time": "09:02"
            },
            "checklist": [
              "Cath lab on standby",
              "ECG on arrival"
            ],
            "disclaimer": "Synthetic patient. Clinician confirms.",
            "generated_at": "2026-10-05T09:03:00Z",
            "model": "gemini-3-flash-preview",
            "summary": "58 year old male, chest pain, BP 85/50."
          }
        ],
        "properties": {
          "atmist": {
            "additionalProperties": true,
            "title": "Atmist",
            "type": "object"
          },
          "checklist": {
            "items": {
              "type": "string"
            },
            "title": "Checklist",
            "type": "array"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "atmist",
          "checklist",
          "summary",
          "disclaimer"
        ],
        "title": "BriefOut",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "The one error body. Some errors add keys: `state`, `fallback`, `retry_after_s`.",
        "examples": [
          {
            "detail": "unknown run",
            "error": "unknown_run"
          }
        ],
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "RunLogEntryOut": {
        "description": "One `runs/{id}/log/{n}` entry as GET /runs/{run_id}/log returns it, oldest first.",
        "examples": [
          {
            "confirmed": true,
            "fields": {
              "complaint": "chest pain",
              "vitals": {
                "dbp": 50,
                "sbp": 85
              }
            },
            "interventions": [],
            "kind": "voice",
            "n": 1,
            "t": "2026-10-05T09:02:00+00:00",
            "transcript_en": "Patient has chest pain, BP 85 over 50"
          }
        ],
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          },
          "interventions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Interventions",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "t": {
            "title": "T",
            "type": "string"
          },
          "transcript_en": {
            "title": "Transcript En",
            "type": "string"
          }
        },
        "required": [
          "n",
          "t",
          "kind",
          "transcript_en",
          "fields",
          "interventions",
          "confirmed"
        ],
        "title": "RunLogEntryOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Preempts traffic signals along an emergency vehicle's route: ticks in, alerts out.\n\n**Errors.** Every 4xx and 5xx body is the same envelope, `{\"error\": \"<code>\", \"detail\": \"<text>\"}`, sometimes with extra keys\n(`state`, `fallback`, `retry_after_s`). Every response carries `X-Request-Id`.\n\n**Device tokens.** Calls that change a run need `X-Device-Token` (handed out by `/vehicles/bind`, `/duty` and\n`/hospital/duty`): 401 without it, 403 with the wrong one.\n\n**Rate limits.** Per IP in one-minute windows shared by every instance: 10 heavy calls (`/triage`, `/log`, `/brief`, `/route`,\n`/cop-note`, `/runs/{run_id}/after-action`), 900 `/location`, 60 everything else. Over the limit: 429 `rate_limited` with\n`Retry-After`.\n\n**Idempotency.** `POST /location` takes an optional `Idempotency-Key` header (or `tick_id` in the body): a repeat within\n10 minutes returns the stored response, with `Idempotent-Replayed: true`, and changes nothing.",
    "title": "Emergency Green Corridor API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/briefs/{run_id}": {
      "get": {
        "description": "The stored ATMIST brief (SCHEMA.md, briefs/{run_id}). Same tokens as the log; 404 `no_brief` until one is written.",
        "operationId": "get_brief_briefs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-device-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Device-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A device token is required (`device_token_required`)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not allowed (`device_token_mismatch`, `unregistered_vehicle`, `run_not_active`, `forbidden`)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not found (`unknown_run`, `unknown_alert`, `unknown_junction`, `not_found`)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed (`validation_error`) or the extraction could not read the input (`extraction_failed`)."
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "The stored hospital brief of a run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/runs/{run_id}/log": {
      "get": {
        "description": "The transit log, oldest first: [{n, t, kind, transcript_en, fields, interventions, confirmed, photo_url?}]. Needs the\nrun's vehicle token or a hospital desk token.",
        "operationId": "run_log_runs__run_id__log_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-device-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Device-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunLogEntryOut"
                  },
                  "title": "Response 200 Run Log Runs  Run Id  Log Get",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A device token is required (`device_token_required`)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not allowed (`device_token_mismatch`, `unregistered_vehicle`, `run_not_active`, `forbidden`)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not found (`unknown_run`, `unknown_alert`, `unknown_junction`, `not_found`)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed (`validation_error`) or the extraction could not read the input (`extraction_failed`)."
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "The transit log of a run",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
...
}